# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank management system
  description: api for managing user's cards
  contact:
    name: Panasik Uladzislau
    url: https://github.com/Vladislav3421730
    email: panasikvladislav1@gmail.com
  version: 1.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/v1/block/{id}/status:
    put:
      tags:
        - Block Requests Management
      summary: Update block request status
      description: Updates the status of an existing block request (e.g., COMPLETED). If COMPLETED, card status will be set to BLOCKED.
      operationId: changeStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockCardRequestDto'
        required: true
      responses:
        '200':
          description: Status was changed successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BlockRequestDto'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/cards:
    get:
      tags:
        - Card Management
      summary: Find all cards
      operationId: findAll
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: minBalance
          in: query
          required: false
          schema:
            type: number
        - name: maxBalance
          in: query
          required: false
          schema:
            type: number
        - name: expiredBefore
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: expiredAfter
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of cards returned successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
    post:
      tags:
        - Card Management
      summary: Add new card
      description: Creates and saves a new card
      operationId: saveCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCardRequestDto'
        required: true
      responses:
        '201':
          description: Card successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Invalid card data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/cards/{id}/status:
    patch:
      tags:
        - Card Management
      summary: Update card status
      description: Changes the status of a card (e.g., ACTIVE, BLOCKED)
      operationId: changeStatus_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusCardRequestDto'
        required: true
      responses:
        '200':
          description: Card status successfully updated
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/users/{id}:
    get:
      tags:
        - Users Management
      summary: Get user by ID
      description: Retrieve user information by their unique identifier
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
    delete:
      tags:
        - Users Management
      summary: Delete user
      description: Deletes a user by their ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User successfully deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/cards/{id}:
    get:
      tags:
        - Card Management
      summary: Get card by ID
      description: Retrieve card information by their unique identifier
      operationId: findById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: card found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
    delete:
      tags:
        - Card Management
      summary: Delete card
      description: Deletes a card by its ID
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card successfully deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/cards/me:
    get:
      tags:
        - Card Management
      summary: Find all cards for authorized user
      operationId: findAllUserCards
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: minBalance
          in: query
          required: false
          schema:
            type: number
        - name: maxBalance
          in: query
          required: false
          schema:
            type: number
        - name: expiredBefore
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: expiredAfter
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of cards returned successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
components:
  schemas:
...
    AppErrorDto:
      type: object
      description: DTO for errors
      properties:
        code:
          type: integer
          format: int32
          description: Error's code
          example: 400
        timestamp:
          type: string
          description: Timestamp of when the error occurred
          example: '2025-02-07 14:30:00'
        message:
          type: string
          description: Error message describing the problem
          example: Invalid request
    BlockCardRequestDto:
      type: object
      description: Request to block a card with a specific status (COMPLETED or REJECTED)
      properties:
        status:
          type: string
          description: The status of the block request. Only COMPLETED or REJECTED are allowed.
          example: COMPLETED
      required:
        - status
    BlockRequestDto:
      type: object
      description: DTO representing a block request for a card, containing status, timestamps, and related IDs.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the block request
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        createdAt:
          type: string
          format: date-time
          description: Date and time when the block request was created
          example: '2025-05-21 15:30:00'
        updatedAt:
          type: string
          format: date-time
          description: Date and time when the block request was last updated
          example: '2025-05-21 16:00:00'
        status:
          type: string
          description: Status of the block request
          enum:
            - COMPLETED
            - REJECTED
          example: COMPLETED
        cardId:
          type: string
          format: uuid
          description: Unique identifier of the card where the transaction was initiated
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        userId:
          type: string
          format: uuid
          description: User ID associated with the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
...
    AddCardRequestDto:
      type: object
      description: Request to add a new card
      properties:
        userId:
          type: string
          format: uuid
          description: ID of the user to whom the card will be linked
          example: 550e8400-e29b-41d4-a716-446655440000
        number:
          type: string
          description: Card number in format XXXX XXXX XXXX XXXX
          example: 1234
      required:
        - number
        - userId
    CardDto:
      type: object
      description: DTO representing a card with details such as number, balance, and expiration date.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        number:
          type: string
          description: Card number
          example: '**** **** **** 3456'
        status:
          type: string
          description: Status of the card
          enum:
            - ACTIVE
            - INACTIVE
            - BLOCKED
          example: ACTIVE
        balance:
          type: number
          description: Current balance on the card
          example: 1000.5
        expirationDate:
          type: string
          format: date
          description: Expiration date of the card
          example: '2025-12-31'
        userId:
          type: string
          format: uuid
          description: User ID associated with the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
...
    StatusCardRequestDto:
      type: object
      description: DTO for updating the card status
      properties:
        status:
          type: string
          description: New card status. The allowed values are defined in the custom @Status annotation
          example: ACTIVE
      required:
        - status
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          $ref: '#/components/schemas/PageMetadata'
    UserDto:
      type: object
      description: DTO representing a user and their associated cards.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        username:
          type: string
          description: User login
          example: user
        firstName:
          type: string
          description: First name of the user
          example: John
        lastName:
          type: string
          description: Last name of the user
          example: Doe
        cards:
          type: array
          description: List of cards associated with the user
          items:
            $ref: '#/components/schemas/CardDto'
        isBan:
          type: boolean
          description: Ban status of the user
          example: false
  securitySchemes:
    BearerAuthentication:
      type: http
      scheme: bearer
      bearerFormat: JWT